requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.19.0
pymysql==1.0.2
mysqlclient==2.1.0
cryptography==36.0.1